Backstretch light theme fix
Backstretch is a script for the slideshow on the dark theme. If it’s
not being used, like on the light theme, the backstretch code in app.js
throws an error and conflicts with other scripts. This code change adds
an if statement that looks to see if the .backstretch class exists and
executes if it does. This fixes 3 issues in one. Now the “home” class
shows up on the light theme home page in the source code.